refactor(libpod): split client methods by HTTP verb (#1390)
## Summary

I split the libpod client methods into focused HTTP verb modules while
keeping the client state, transport plumbing, and status handling in
`internal/libpod/client/mod.rs`. The public `Client` API remains
unchanged and every production file stays below the repository's hard
line limit.

## Validation

- `cargo build --locked --bin podup --features test-helpers`
- `cargo test --locked --lib --features test-helpers`
- `cargo test --locked --bins --features test-helpers`
- `rustfmt --check --edition 2021 internal/libpod/client/mod.rs
internal/libpod/client/get.rs internal/libpod/client/post.rs
internal/libpod/client/put.rs internal/libpod/client/misc.rs
internal/libpod/client/delete.rs`

The real-Podman integration lane remains the final runtime check for
both supported Podman majors.

Closes #1386
